refactor: replace address with cp location
