+++ core/fixtures/todos__validate_screen_cplocation_admin.json
@@ -1,0 +1,10 @@
+[
+  {
+    "model": "core.todo",
+    "pk": 12,
+    "fields": {
+      "description": "Validate screen CPLocation admin",
+      "url": "/admin/ocpp/cplocation/"
+    }
+  }
+]
